laser_device/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
PyAudio==0.2.13
pygame==2.1.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
scipy==1.10.1
six==1.16.0
tqdm==4.64.1
